pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "psycopg2-client"
version = "1.0.3"
description = "Psycopg2 helper function to run PostgreSQL query with #if support"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Gu Park", email = "doctorgu@kakao.com"}
]
keywords = ["psycopg2 client", "psycopg2 helper"]
dependencies = [
    "psycopg2-binary",
    "pyhumps",
    "typing-extensions",
    "python-dotenv",
    "pydantic-settings",
    "Flask",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-mock",
    "pytest-asyncio",
]

[project.urls]
homepage = "https://github.com/doctorgu/psycopg2-client"
repository = "https://github.com/doctorgu/psycopg2-client"
documentation = "https://github.com/doctorgu/psycopg2-client"
